requirements.txt
bottle==0.12.25
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
joblib==1.3.2
numpy==1.26.4
pandas==2.2.1
Paste==3.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
RPi.GPIO==0.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
threadpoolctl==3.4.0
tzdata==2024.1
urllib3==2.2.1
